Implement Magentic-One Orchestrator
