+++ magda_agent/agents/__init__.py
@@ -1,5 +1,4 @@
-from .magentic_one import MagenticOneOrchestrator
 from .teams import TeamManager
 from .team_orchestrator import TeamOrchestrator
 
-__all__ = ["TeamManager", "MagenticOneOrchestrator", "TeamOrchestrator"]
+__all__ = ["TeamManager", "TeamOrchestrator"]
