+++ magda_agent/architecture/__init__.py
@@ -1,0 +1,3 @@
+from .magentic_one import MagenticOneOrchestrator
+
+__all__ = ["MagenticOneOrchestrator"]
